fix: align synthetic provider handling (#171) (thanks @monotykamary)
