+++ Sources/CodexBarCLI/TokenAccountCLI.swift
@@ -137,7 +137,7 @@ struct TokenAccountCLIContext {
                 augment: ProviderSettingsSnapshot.AugmentProviderSettings(
                     cookieSource: .manual,
                     manualCookieHeader: header))
-        case .codex, .gemini, .antigravity, .zai, .copilot, .kiro, .vertexai, .kimi, .kimik2, .amp:
+        case .codex, .gemini, .antigravity, .zai, .copilot, .kiro, .vertexai, .kimi, .kimik2, .amp, .synthetic:
             return nil
         }
     }
